+++ src/config.js
@@ -18,6 +18,8 @@ const defaultConfig = {
   "status": "Message me for help!",
   "responseMessage": "Thank you for your message! Our mod team will reply to you here as soon as possible.",
 
+  "newThreadCategoryId": null,
+
   "inboxServerPermission": null,
   "alwaysReply": false,
   "alwaysReplyAnon": false,