+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
+## Sep 22, 2017
+* Added `newThreadCategoryId` option. This option can be set to a category ID to place all new threads in that category.
+
 ## Sep 20, 2017
 * Fixed crash when the bot was unable to find or create a modmail thread
 * Reduced error log spam in case of network errors from Eris
-* Fix unintended error when a message was ignored due to an "accidental thread" word
+* Fixed unintended error when a message was ignored due to an "accidental thread" word
 
 ## Sep 19, 2017
 * Added `logChannelId` option